+++ semver-core/common/main/kotlin/com/javiersc/semver/VersionFormat.kt
@@ -4,17 +4,17 @@ import kotlin.text.RegexOption.IGNORE_CASE
 
 /**
  * Samples:
- *
- * - Normal: `1.0.0-alpha.01`
+ * - Default: `1.0.0-alpha.01`
  */
 internal enum class VersionFormat(val scope: Regex, val stage: Regex) {
     Default(
-        """^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)""".toRegex(),
-        """(?!SNAPSHOT\.\d)([a-zA-Z]+(\.\d+)|\bSNAPSHOT\b)""".toRegex(IGNORE_CASE)
+        Regex("""^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)"""),
+        Regex("""(?!SNAPSHOT\.\d)([a-zA-Z]+(\.\d+)|\bSNAPSHOT\b)""", IGNORE_CASE)
     );
 
     val regex: Regex =
-        """^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(-(?!SNAPSHOT\.\d)([a-zA-Z]+(\.\d+)|\bSNAPSHOT\b))?$""".toRegex(
+        Regex(
+            """^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(-(?!SNAPSHOT\.\d)([a-zA-Z]+(\.\d+)|\bSNAPSHOT\b))?$""",
             IGNORE_CASE
         )
 }